+++ Src/Pscx/Commands/EnvironmentBlock/GetPathVariableCommand.cs
@@ -12,6 +12,7 @@ using System.Management.Automation;
 
 namespace Pscx.Commands.EnvironmentBlock
 {
+    [OutputType(typeof(string))]
     [Cmdlet(VerbsCommon.Get, PscxNouns.PathVariable), Description("Gets the specified path-like environment variable, defaults to PATH")]
     [RelatedLink(typeof(AddPathVariableCommand))]
     [RelatedLink(typeof(SetPathVariableCommand))]